Rename AnalyticsTabsContainer
