+++ client/src/store/sagas.js
@@ -1,6 +1,6 @@
 import { all } from 'redux-saga/effects';
 import authSaga from '../containers/LoginPageContainer/sagas';
-import analyticsSaga from '../containers/AnalyticsTabsContainer/sagas';
+import analyticsSaga from '../containers/AnalyticsContainer/sagas';
 import currentDashboardSaga from '../containers/DashboardContainer/sagas';
 import currentVisualizationSaga from '../containers/ViewVisualizationContainer/sagas';
 import usersSaga from '../containers/PeoplePageContainer/sagas';
